Fix hover effect in JumpToLetter view.
If the button is very wide like in the app list, the hover radius should
stay width/2 so that hovering the center show a line above and below, but
doesn't highlight the left/right side of the button.

If we use the same effect radius in a square, hovering the dead center
will not highlight any of the edges. In this case, we use height as the
radius.

Note that this doesn't account for very tall buttons, like the tile grid,
but we manually override the radius in that case.