Override the default ConfigurationShortcuts view to add info text.
There's no documented way for the user to know that Alt+F1 == Meta
without googling it. When a user adds this widget, they might not use
"Alternatives" and instead add a new widget and remove the old one.
When they do it like that, the global shortcut is not setup. Thus
neither Alt+F1, nor Meta key will open the menu.

Unfortunately, there isn't really a way to detect is the global
shortcut is set as QKeySequence is not QML friendly.