Don't fetch linked file contents it's not a .desktop in tile editor
This was probably checked outside this function before a refactor of
the preset tiles. Luckily we haven't had any reports of increased RAM
when editing tiles yet, but it's stands to reason a that if a user
linked to a huge file, it RAM would spike opening the tile editor.