Configurable sidebar width and icon size.
Neeeded to bind sidebarIconSize.maximumValue to configValue in the
config. Binding to configSpinBox.value has a race condition, and it
changed the icon size to the minimum 16px before it fully loaded the
configValue from the config.

This affects all FlatButtons including the sidebar submenu popup and
the search filter view buttons. I'm undecided on wither I should make
the sizes a different property.