Simple demo implementation for auto coloring tiles (Issue #33)
We'll need to autogenerate the text color as well.

The app list "tiles" would need spacing between the rectangles. This
means the icons will need to be smaller or the buttons will be taller
than the icons so there is space for the button padding.

The bg color that is currently generated is fairly bright, so we'll
probably need to make it a litter darker... or somehow make it depend
on the theme bg color.