Add LibConfig.SpinBox based on QQC2.SpinBox
Since QQC2.SpinBox is terrible compared to QQC1, we need reimplement
many features like real values with decimals and prefix/suffix.

Since QQC2 uses an integer for the value property, that limits the
size of a real number to the range of a signed int. A real with 2
decimals will reset to 0 when trying to enter a number above 11mil.

I first placed the prefix/suffix in the SpinBox text, however
filtering it out was an issue. I ended up drawing them in separate
Labels which gives us the ability to color their text.

To make room for the prefix/suffix Labels, we bind the TextInput's
leftPadding/rightPadding instead of the SpinBox's so that clicking on
the prefix/suffix will focus the TextInput.

Since the KDE Desktop QQC2 theme formats the text after each keypress
it would move the cursor to the end of the string after replacing the
TextInput text. Typing '44' with decimals=2 would turn into '4.004'.
To fix this, we scan the SpinBox for a Connections and destroy it.
We then bind to TextInput.onTextEdited ourself and keep track of where
the cursor and period is. After the TextInput text is replaced, we
move the cursor to the correct position.

When focused, we select all like QQC1 does.