Add MININT and MAXINT to SpinBox as properties. Fix minus handling.
Fix a few issues when typing '-'. When the text was just '-' or it
contained '-.' the valueFromText() function would contain a NaN value.
The SpinBox value would then equal the min value (which was MININT
when I was testing). When an NaN is detected, the value returns to 0.
This makes it harder to type '-5' as you first need to type '5' then
add the '-' in front. If you type '-', valueFromText() == 0 then
textFromValue() replaces the text with '0', which deletes the '-' you
just typed.

We also move the minus sign to the start of the text.

I added back the debug logging commented out since I expect more
issues to appear.