Add LibConfig.RadioButtonGroup for smaller spacing
We automatically set Kirigami.FormData.buddyFor to the first button so
that the Kirigami label is lined up correctly.

We also create a QQC2.ButtonGroup just in case you wrap a RadioButton
in a Layout. By default in QQC2, a RadioButton will automatically
group with sibling RadioButtons in the same parent.